Refactor staking hooks to simplify allowance handling
