+++ src/app/user/Stake/hooks/useStakeRIF.tsx
@@ -1,102 +1,27 @@
 import { ActionHookToUse } from '@/app/user/Stake/StakingContext'
-import { RIFTokenAbi } from '@/lib/abis/RIFTokenAbi'
 import { StRIFTokenAbi } from '@/lib/abis/StRIFTokenAbi'
-import { tokenContracts } from '@/lib/contracts'
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { Address, Hash, parseEther } from 'viem'
-import { useAccount, useReadContract, useWaitForTransactionReceipt, useWriteContract } from 'wagmi'
-import { CustomStakingRIFFooter } from '../CustomStakingRIFFooter'
-import { useTxStatusContext } from '@/shared/context/TxStatusContext'
+import { useCallback } from 'react'
+import { Address, parseEther } from 'viem'
+import { useAccount, useWriteContract } from 'wagmi'
 
-export const useStakeRIF: ActionHookToUse = (
-  amount: string,
-  tokenToSendContract: string,
-  tokenToReceiveContract: string,
-) => {
+export const useStakeRIF: ActionHookToUse = (amount: string, tokenToReceiveContract: string) => {
   const { address } = useAccount()
-  const [allowanceHash, setAllowanceHashUsed] = useState<Hash>()
-  const { trackTransaction } = useTxStatusContext()
-
-  const { data: allowanceBalance, isLoading: isAllowanceReadLoading } = useReadContract({
-    abi: RIFTokenAbi,
-    address: tokenContracts.RIF,
-    functionName: 'allowance',
-    args: [address!, tokenToReceiveContract as Address],
-    query: {
-      refetchInterval: 5000,
-    },
-  })
-  const isAllowanceEnough = useMemo(
-    () => !!(allowanceBalance && allowanceBalance >= parseEther(amount)),
-    [amount, allowanceBalance],
-  )
-
-  const {
-    writeContractAsync: requestAllowance,
-    data: allowanceTxHash,
-    isPending: isRequestingAllowance,
-  } = useWriteContract()
-
-  const tx = useWaitForTransactionReceipt({
-    hash: allowanceHash,
-  })
-
-  const { isPending: isAllowanceTxPending, failureReason: isAllowanceTxFailed } = tx
-
-  useEffect(() => {
-    if (allowanceTxHash) {
-      setAllowanceHashUsed(allowanceTxHash)
-      trackTransaction(allowanceTxHash)
-    }
-  }, [allowanceTxHash, trackTransaction])
-
-  const onRequestAllowance = useCallback(
-    () =>
-      requestAllowance(
-        {
-          abi: RIFTokenAbi,
-          address: tokenToSendContract as Address,
-          functionName: 'approve',
-          args: [tokenToReceiveContract as Address, parseEther(amount)],
-        },
-        {
-          onSuccess: txHash => setAllowanceHashUsed(txHash),
-        },
-      ),
-    [amount, requestAllowance, tokenToReceiveContract, tokenToSendContract],
-  )
   const { writeContractAsync: stake, isPending } = useWriteContract()
 
-  const onRequestStake = () =>
-    stake({
-      abi: StRIFTokenAbi,
-      address: tokenToReceiveContract as Address,
-      functionName: 'depositAndDelegate',
-      args: [address!, parseEther(amount)],
-    })
-
-  const customFooter = useMemo(
-    () => (
-      <CustomStakingRIFFooter
-        hash={allowanceHash}
-        isAllowanceNeeded={!isAllowanceEnough}
-        isAllowanceTxPending={allowanceHash && isAllowanceTxPending && !isAllowanceTxFailed}
-        isAllowanceReadLoading={isAllowanceReadLoading}
-        isAllowanceTxFailed={!!isAllowanceTxFailed}
-      />
-    ),
-    [allowanceHash, isAllowanceEnough, isAllowanceTxPending, isAllowanceReadLoading, isAllowanceTxFailed],
+  const onRequestStake = useCallback(
+    () =>
+      stake({
+        abi: StRIFTokenAbi,
+        address: tokenToReceiveContract as Address,
+        functionName: 'depositAndDelegate',
+        args: [address!, parseEther(amount)],
+      }),
+    [address, amount, stake, tokenToReceiveContract],
   )
+
   return {
-    isAllowanceEnough,
     onConfirm: onRequestStake,
-    customFooter,
-    isAllowanceReadLoading,
+    customFooter: null,
     isPending,
-    onRequestAllowance,
-    isRequestingAllowance,
-    isAllowanceTxPending: allowanceHash && isAllowanceTxPending && !isAllowanceTxFailed,
-    isAllowanceTxFailed: allowanceHash && isAllowanceTxFailed,
-    allowanceHash,
   }
 }
